requirements.txt
# Cloud Provider SDKs
boto3>=1.34.0
google-cloud-billing>=1.12.0
google-cloud-bigquery>=3.14.0

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23

# Configuration and utilities
python-dotenv>=1.0.0

# Optional: for better logging
colorlog>=6.8.0
